+++ CHANGELOG.md
@@ -21,11 +21,14 @@
 
 ## v1.6.4
 
-**Improvements/Fixes**
+**New Features**
+- Add a new function `quill::get_all_loggers()` that returns all the existing loggers ([##114](https://github.com/odygrd/quill/issues/#114))
+
+- **Improvements/Fixes**
 
 - Fix initializations for C++17
 - Fix compiler warning in check_format function.
-- Replace QUILL_DUAL_QUEUE_MODE with QUILL_DISABLE_DUAL_QUEUE_MODE
+- Replace `QUILL_DUAL_QUEUE_MODE` with `QUILL_DISABLE_DUAL_QUEUE_MODE`
 - Update bundled fmt to 8.1.1
 - Minor performance and accuracy improvements to rdtsc clock used by the backend thread.
 - Fix compile error when C++20 is used. ([#162](https://github.com/odygrd/quill/issues/162))